Cleanup address slice to fix circular import
