+++ apps/mobile-wallet/src/components/AddressFlashListScreen.tsx
@@ -3,7 +3,7 @@ import { AddressHash } from '@alephium/shared'
 import AddressBox from '~/components/AddressBox'
 import FlashListScreen, { FlashListScreenProps } from '~/components/layout/FlashListScreen'
 import { useAppSelector } from '~/hooks/redux'
-import { selectAllAddresses } from '~/store/addressesSlice'
+import { selectAllAddresses } from '~/store/addresses/addressesSelectors'
 import { DEFAULT_MARGIN } from '~/style/globalStyle'
 import { Address } from '~/types/addresses'
