+++ apps/mobile-wallet/src/components/RefreshSpinner.tsx
@@ -3,7 +3,7 @@ import { RefreshControl, RefreshControlProps } from 'react-native'
 import { useTheme } from 'styled-components/native'
 
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
-import { syncLatestTransactions } from '~/store/addressesSlice'
+import { syncLatestTransactions } from '~/store/addresses/addressesActions'
 
 const RefreshSpinner = (props: Partial<RefreshControlProps>) => {
   const theme = useTheme()
