+++ apps/mobile-wallet/src/components/layout/TransactionsFlashList.tsx
@@ -10,7 +10,7 @@ import EmptyPlaceholder from '~/components/EmptyPlaceholder'
 import RefreshSpinner from '~/components/RefreshSpinner'
 import { openModal } from '~/features/modals/modalActions'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
-import { syncAllAddressesTransactionsNextPage } from '~/store/addressesSlice'
+import { syncAllAddressesTransactionsNextPage } from '~/store/addresses/addressesActions'
 import { DEFAULT_MARGIN } from '~/style/globalStyle'
 import { AddressConfirmedTransaction, AddressPendingTransaction, AddressTransaction } from '~/types/transactions'
 import { isPendingTx } from '~/utils/transactions'
