+++ apps/mobile-wallet/src/contexts/walletConnect/WalletConnectSessionProposalModal.tsx
@@ -28,8 +28,9 @@ import { persistSettings } from '~/features/settings/settingsPersistentStorage'
 import usePersistAddressSettings from '~/hooks/layout/usePersistAddressSettings'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import { getAddressAsymetricKey, initializeKeyringWithStoredWallet } from '~/persistent-storage/wallet'
-import { selectAddressesInGroup } from '~/store/addresses/addressesSelectors'
-import { newAddressGenerated, selectAllAddresses, syncLatestTransactions } from '~/store/addressesSlice'
+import { syncLatestTransactions } from '~/store/addresses/addressesActions'
+import { selectAddressesInGroup, selectAllAddresses } from '~/store/addresses/addressesSelectors'
+import { newAddressGenerated } from '~/store/addressesSlice'
 import { VERTICAL_GAP } from '~/style/globalStyle'
 import { Address } from '~/types/addresses'
 import { getRandomLabelColor } from '~/utils/colors'
