+++ apps/mobile-wallet/src/features/addressesManagement/AddressSettingsModal.tsx
@@ -18,7 +18,8 @@ import withModal from '~/features/modals/withModal'
 import usePersistAddressSettings from '~/hooks/layout/usePersistAddressSettings'
 import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import AddressForm, { AddressFormData } from '~/screens/Addresses/Address/AddressForm'
-import { addressSettingsSaved, selectAddressByHash } from '~/store/addressesSlice'
+import { selectAddressByHash } from '~/store/addresses/addressesSelectors'
+import { addressSettingsSaved } from '~/store/addressesSlice'
 import { copyAddressToClipboard } from '~/utils/addresses'
 import { showExceptionToast } from '~/utils/layout'
