+++ apps/mobile-wallet/src/features/addressesManagement/useCanDeleteAddress.tsx
@@ -1,7 +1,7 @@
 import { AddressHash } from '@alephium/shared'
 
 import { useAppSelector } from '~/hooks/redux'
-import { selectAddressByHash } from '~/store/addressesSlice'
+import { selectAddressByHash } from '~/store/addresses/addressesSelectors'
 
 const useCanDeleteAddress = (addressHash: AddressHash) => {
   const address = useAppSelector((s) => selectAddressByHash(s, addressHash))
