+++ apps/mobile-wallet/src/features/assetsDisplay/tokenDisplay/tokenDetailsModal/TokenDetailsModalHeader.tsx
@@ -8,7 +8,7 @@ import AssetLogo from '~/components/AssetLogo'
 import Badge from '~/components/Badge'
 import { TokenDetailsModalCommonProps } from '~/features/assetsDisplay/tokenDisplay/tokenDetailsModal/tokenDetailsModalTypes'
 import { useAppSelector } from '~/hooks/redux'
-import { makeSelectAddressesKnownFungibleTokens } from '~/store/addressesSlice'
+import { makeSelectAddressesKnownFungibleTokens } from '~/store/addresses/addressesSelectors'
 
 const TokenDetailsModalHeader = ({ tokenId, addressHash }: TokenDetailsModalCommonProps) => {
   const selectAddressesKnownFungibleTokens = useMemo(makeSelectAddressesKnownFungibleTokens, [])
