+++ apps/mobile-wallet/src/features/receive/ReceiveQRCodeSection.tsx
@@ -10,7 +10,7 @@ import { ScreenSection } from '~/components/layout/Screen'
 import Surface from '~/components/layout/Surface'
 import Row from '~/components/Row'
 import { useAppSelector } from '~/hooks/redux'
-import { selectAddressByHash } from '~/store/addressesSlice'
+import { selectAddressByHash } from '~/store/addresses/addressesSelectors'
 import { BORDER_RADIUS_BIG } from '~/style/globalStyle'
 import { copyAddressToClipboard } from '~/utils/addresses'
