+++ apps/mobile-wallet/src/screens/Addresses/Address/NewAddressScreen.tsx
@@ -12,7 +12,9 @@ import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import RootStackParamList from '~/navigation/rootStackRoutes'
 import { initializeKeyringWithStoredWallet } from '~/persistent-storage/wallet'
 import AddressForm, { AddressFormData } from '~/screens/Addresses/Address/AddressForm'
-import { newAddressGenerated, selectAllAddresses, syncLatestTransactions } from '~/store/addressesSlice'
+import { syncLatestTransactions } from '~/store/addresses/addressesActions'
+import { selectAllAddresses } from '~/store/addresses/addressesSelectors'
+import { newAddressGenerated } from '~/store/addressesSlice'
 import { getRandomLabelColor } from '~/utils/colors'
 import { showExceptionToast } from '~/utils/layout'
