+++ apps/mobile-wallet/src/store/addresses/addressesStorageUtils.ts
@@ -2,7 +2,8 @@ import { keyring } from '@alephium/keyring'
 import { AddressMetadata } from '@alephium/shared'
 
 import { initializeKeyringWithStoredWallet } from '~/persistent-storage/wallet'
-import { newAddressGenerated, syncLatestTransactions } from '~/store/addressesSlice'
+import { syncLatestTransactions } from '~/store/addresses/addressesActions'
+import { newAddressGenerated } from '~/store/addressesSlice'
 import { store } from '~/store/store'
 import { WalletMetadata } from '~/types/wallet'
 import { persistAddressesSettings } from '~/utils/addresses'
