+++ apps/mobile-wallet/src/store/addressesSlice.ts
@@ -1,40 +1,23 @@
 import {
-  AddressFungibleToken,
-  AddressHash,
   appReset,
-  Asset,
   balanceHistoryAdapter,
-  calculateAssetsData,
   customNetworkSettingsSaved,
   extractNewTransactions,
   getTransactionsOfAddress,
-  networkPresetSwitched,
-  NFT,
-  selectAllFungibleTokens,
-  selectAllNFTs,
-  selectAllPrices,
-  selectNFTIds,
-  sortAssets,
-  TokenDisplayBalances
+  networkPresetSwitched
 } from '@alephium/shared'
-import { ALPH } from '@alephium/token-list'
-import { explorer, groupOfAddress } from '@alephium/web3'
-import { Transaction } from '@alephium/web3/dist/src/api/api-explorer'
-import {
-  createAsyncThunk,
-  createEntityAdapter,
-  createSelector,
-  createSlice,
-  EntityState,
-  isAnyOf,
-  PayloadAction
-} from '@reduxjs/toolkit'
+import { groupOfAddress } from '@alephium/web3'
+import { createSlice, EntityState, isAnyOf, PayloadAction } from '@reduxjs/toolkit'
 
-import { fetchAddressesBalances, fetchAddressesTokens, fetchAddressesTransactionsPage } from '~/api/addresses'
-import { selectHiddenAssetsIds } from '~/features/assetsDisplay/hideAssets/hiddenAssetsSelectors'
 import { addressMetadataIncludesHash } from '~/persistent-storage/wallet'
-import { addressDeleted } from '~/store/addresses/addressesActions'
-import { RootState } from '~/store/store'
+import {
+  addressDeleted,
+  syncAddressesBalances,
+  syncAddressesTokens,
+  syncAllAddressesTransactionsNextPage,
+  syncLatestTransactions
+} from '~/store/addresses/addressesActions'
+import { addressesAdapter } from '~/store/addresses/addressesAdaptor'
 import {
   appLaunchedWithLastUsedWallet,
   newWalletGenerated,
@@ -46,16 +29,6 @@ import { getRandomLabelColor } from '~/utils/colors'
 
 const sliceName = 'addresses'
 
-const addressesAdapter = createEntityAdapter<Address>({
-  selectId: (address) => address.hash,
-  sortComparer: (a, b) => {
-    // Always keep main address to the top of the list
-    if (a.settings.isDefault) return -1
-    if (b.settings.isDefault) return 1
-    return (b.lastUsed ?? 0) - (a.lastUsed ?? 0)
-  }
-})
-
 interface AddressesState extends EntityState<Address> {
   status: 'uninitialized' | 'initialized'
   balancesStatus: 'uninitialized' | 'initialized'
@@ -65,109 +38,6 @@ const initialState: AddressesState = addressesAdapter.getInitialState({
   status: 'uninitialized',
   balancesStatus: 'uninitialized'
 })
-
-export const syncLatestTransactions = createAsyncThunk(
-  'addresses/syncLatestTransactions',
-  async (
-    payload: { addresses: AddressHash | AddressHash[] | 'all'; areAddressesNew: boolean },
-    { getState, dispatch }
-  ) => {
-    const { addresses: _addresses, areAddressesNew } = payload
-    const state = getState() as RootState
-
-    const addresses =
-      _addresses === 'all'
-        ? (state.addresses.ids as AddressHash[])
-        : Array.isArray(_addresses)
-          ? _addresses
-          : [_addresses]
-
-    if (areAddressesNew)
-      Promise.all([dispatch(syncAddressesBalances(addresses)), dispatch(syncAddressesTokens(addresses))])
-
-    const latestTransactions = await fetchAddressesTransactionsPage(addresses, 1)
-
-    const newTransactionsResults = addresses.reduce(
-      (acc, addressHash) => {
-        const transactionsOfAddress = getTransactionsOfAddress(latestTransactions, addressHash)
-        const address = selectAddressByHash(state, addressHash)
-
-        if (!address) return acc
-
-        const newTransactions = extractNewTransactions(transactionsOfAddress, address.transactions)
-
-        if (newTransactions.length > 0)
-          acc.push({
-            hash: address.hash,
-            newTransactions
-          })
-
-        return acc
-      },
-      [] as {
-        hash: AddressHash
-        newTransactions: Transaction[]
-      }[]
-    )
-
-    const addressesWithNewTransactions = newTransactionsResults.map(({ hash }) => hash)
-    const addressesToFetchData =
-      state.addresses.status === 'uninitialized' ? (state.addresses.ids as AddressHash[]) : addressesWithNewTransactions
-
-    if (!areAddressesNew && addressesToFetchData.length > 0)
-      Promise.all([
-        dispatch(syncAddressesBalances(addressesToFetchData)),
-        dispatch(syncAddressesTokens(addressesToFetchData))
-      ])
-
-    return newTransactionsResults
-  }
-)
-
-export const syncAddressesBalances = createAsyncThunk(
-  `${sliceName}/syncAddressesBalances`,
-  async (addresses: AddressHash[]) => await fetchAddressesBalances(addresses)
-)
-
-export const syncAddressesTokens = createAsyncThunk(
-  `${sliceName}/syncAddressesTokens`,
-  async (addresses: AddressHash[]) => await fetchAddressesTokens(addresses)
-)
-
-export const syncAllAddressesTransactionsNextPage = createAsyncThunk(
-  'addresses/syncAllAddressesTransactionsNextPage',
-  async (
-    payload: { minTxs: number } | undefined,
-    { getState }
-  ): Promise<{ pageLoaded: number; transactions: explorer.Transaction[] }> => {
-    const state = getState() as RootState
-    const addresses = selectAllAddresses(state)
-    const addressesHashes = addresses.map(({ hash }) => hash)
-    const minimumNewTransactionsNeeded = payload?.minTxs ?? 1
-
-    let nextPageToLoad = state.confirmedTransactions.pageLoaded + 1
-    let enoughNewTransactionsFound = false
-    let newTransactions: explorer.Transaction[] = []
-
-    while (!enoughNewTransactionsFound) {
-      const nextPageTransactions = await fetchAddressesTransactionsPage(addressesHashes, nextPageToLoad)
-
-      if (nextPageTransactions.length === 0) break
-
-      newTransactions = newTransactions.concat(
-        nextPageTransactions.filter(
-          (newTx) =>
-            !addresses.some((address) => address.transactions.some((existingTxHash) => existingTxHash === newTx.hash))
-        )
-      )
-
-      enoughNewTransactionsFound = newTransactions.length >= minimumNewTransactionsNeeded
-      nextPageToLoad += 1
-    }
-
-    return { pageLoaded: nextPageToLoad - 1, transactions: newTransactions }
-  }
-)
 
 const addressesSlice = createSlice({
   name: sliceName,
@@ -302,126 +172,6 @@ const addressesSlice = createSlice({
   }
 })
 
-export const {
-  selectById: selectAddressByHash,
-  selectAll: selectAllAddresses,
-  selectIds: selectAddressIds
-} = addressesAdapter.getSelectors<RootState>((state) => state[sliceName])
-
-// Same as in desktop wallet
-export const makeSelectAddressesAlphAsset = () =>
-  createSelector(makeSelectAddresses(), (addresses): Asset => {
-    const alphBalances = addresses.reduce(
-      (acc, { balance, lockedBalance }) => ({
-        balance: acc.balance + BigInt(balance),
-        lockedBalance: acc.lockedBalance + BigInt(lockedBalance)
-      }),
-      { balance: BigInt(0), lockedBalance: BigInt(0) }
-    )
-
-    return {
-      ...ALPH,
-      ...alphBalances,
-      verified: true
-    }
-  })
-
-export const makeSelectAddressesTokens = () =>
-  createSelector(
-    [
-      selectAllFungibleTokens,
-      selectAllNFTs,
-      makeSelectAddressesAlphAsset(),
-      makeSelectAddresses(),
-      selectAllPrices,
-      (_state: RootState, _addressHash?: AddressHash | AddressHash[], filterHiddenTokens?: boolean) =>
-        filterHiddenTokens ?? false,
-      selectHiddenAssetsIds
-    ],
-    (fungibleTokens, nfts, alphAsset, addresses, tokenPrices, filterHiddenTokens, hiddenAssetIds): Asset[] => {
-      const tokenBalances = getAddressesTokenBalances(addresses)
-
-      if (alphAsset.balance > BigInt(0)) {
-        tokenBalances.push(alphAsset)
-      }
-
-      const tokens = calculateAssetsData(tokenBalances, fungibleTokens, nfts, tokenPrices)
-
-      return sortAssets(filterHiddenTokens ? tokens.filter((t) => !hiddenAssetIds.includes(t.id)) : tokens)
-    }
-  )
-
-// Same as in desktop wallet
-export const makeSelectAddressesKnownFungibleTokens = () =>
-  createSelector([makeSelectAddressesTokens(), selectHiddenAssetsIds], (tokens): AddressFungibleToken[] =>
-    tokens.filter((token): token is AddressFungibleToken => !!token.symbol)
-  )
-
-export const makeSelectAddressesUnknownTokens = () =>
-  createSelector([makeSelectAddressesTokens()], (tokens): Asset[] =>
-    tokens.filter((token): token is Asset => !token.name && !token.symbol && !token.verified && !token.logoURI)
-  )
-
-// Same as in desktop wallet
-export const makeSelectAddressesUnknownTokensIds = () =>
-  createSelector(
-    [selectAllFungibleTokens, selectNFTIds, makeSelectAddresses()],
-    (fungibleTokens, nftIds, addresses): Asset['id'][] => {
-      const tokensWithoutMetadata = getAddressesTokenBalances(addresses).reduce(
-        (acc, token) => {
-          const hasTokenMetadata = !!fungibleTokens.find((t) => t.id === token.id)
-          const hasNFTMetadata = nftIds.includes(token.id)
-
-          if (!hasTokenMetadata && !hasNFTMetadata) {
-            acc.push(token.id)
-          }
-
-          return acc
-        },
-        [] as Asset['id'][]
-      )
-
-      return tokensWithoutMetadata
-    }
-  )
-
-// Same as in desktop wallet
-export const makeSelectAddressesCheckedUnknownTokens = () =>
-  createSelector(
-    [makeSelectAddressesUnknownTokensIds(), (state: RootState) => state.app.checkedUnknownTokenIds],
-    (tokensWithoutMetadata, checkedUnknownTokenIds) =>
-      tokensWithoutMetadata.filter((tokenId) => checkedUnknownTokenIds.includes(tokenId))
-  )
-
-// Same as in desktop wallet
-export const makeSelectAddressesNFTs = () =>
-  createSelector([selectAllNFTs, makeSelectAddresses()], (nfts, addresses): NFT[] => {
-    const addressesTokenIds = addresses.flatMap(({ tokens }) => tokens.map(({ tokenId }) => tokenId))
-
-    return nfts.filter((nft) => addressesTokenIds.includes(nft.id))
-  })
-
-// Same as in desktop wallet
-export const makeSelectAddresses = () =>
-  createSelector(
-    [selectAllAddresses, (_, addressHashes?: AddressHash[] | AddressHash) => addressHashes],
-    (allAddresses, addressHashes) =>
-      addressHashes
-        ? allAddresses.filter((address) =>
-            Array.isArray(addressHashes) ? addressHashes.includes(address.hash) : addressHashes === address.hash
-          )
-        : allAddresses
-  )
-
-export const selectDefaultAddress = createSelector(
-  selectAllAddresses,
-  (addresses) => addresses.find((address) => address.settings.isDefault) || addresses[0]
-)
-
-export const selectTotalBalance = createSelector([selectAllAddresses], (addresses) =>
-  addresses.reduce((acc, address) => acc + BigInt(address.balance), BigInt(0))
-)
-
 export const { newAddressGenerated, addressesImported, addressSettingsSaved } = addressesSlice.actions
 
 export default addressesSlice
@@ -465,34 +215,3 @@ const clearAddressesNetworkData = (state: AddressesState) => {
 
   state.status = 'uninitialized'
 }
-
-// Same as in desktop wallet
-const getAddressesTokenBalances = (addresses: Address[]) =>
-  addresses.reduce((acc, { tokens }) => {
-    tokens.forEach((token) => {
-      const existingToken = acc.find((t) => t.id === token.tokenId)
-
-      if (!existingToken) {
-        acc.push({
-          id: token.tokenId,
-          balance: BigInt(token.balance),
-          lockedBalance: BigInt(token.lockedBalance)
-        })
-      } else {
-        existingToken.balance = existingToken.balance + BigInt(token.balance)
-        existingToken.lockedBalance = existingToken.lockedBalance + BigInt(token.lockedBalance)
-      }
-    })
-
-    return acc
-  }, [] as TokenDisplayBalances[])
-
-export const makeSelectAddressesVerifiedFungibleTokens = () =>
-  createSelector([makeSelectAddressesTokens()], (tokens): AddressFungibleToken[] =>
-    tokens.filter((token): token is AddressFungibleToken => !!token.verified)
-  )
-
-export const selectAllAddressVerifiedFungibleTokenSymbols = createSelector(
-  makeSelectAddressesVerifiedFungibleTokens(),
-  (verifiedFungibleTokens) => verifiedFungibleTokens.map((token) => token.symbol)
-)
